Swap GORM to Postgres with owned schema, hooks, and atomic upserts (Refs #16) (#32)
* Swap GORM to Postgres with owned schema, hooks, and atomic upserts

Move OK Folio onto its own dedicated Postgres instead of the legacy MariaDB.

Driver/config:
- Swap gorm.io/driver/mysql -> gorm.io/driver/postgres (pgx stdlib); drop the
  MySQL driver, keep sqlite for unit tests.
- Rewrite DSN() to key/value form with TimeZone=UTC; DATABASE_URL returned
  verbatim. Add DB_PORT/DB_SSLMODE/DATABASE_URL env parsing and Port/SSLMode/URL
  on the config struct; defaults point at the postgres service on 5432, sslmode
  disable.
- Boot guard: refuse a legacy DB_HOST, assert the postgres dialect, and reject a
  legacy MySQL DSN before gorm.Open (testguard); CI lint blocks the mysql driver
  in production code.

Schema/models:
- uint64 ids; text/timestamptz/bytea mapping; PK promoted to GENERATED BY
  DEFAULT AS IDENTITY via idempotent post-migrate Exec so the loader can insert
  legacy ids.
- url_hash bytea NOT NULL (unique) populated by a single BeforeSave hook; raw url
  carries no btree. Re-key IsPhotoDownloaded/RecordDownload/MarkPhotoFailed/
  RecordFailedDownload to url_hash; drop url from the search predicate.
- Atomic upserts via clause.OnConflict; convert-loser-to-Inbox on 23505.
- Net-new columns: favorite, provider, category (derived, replaces the N+1
  filter), content_hash, perceptual_hash (bigint); embedding vector(512) added
  via guarded post-migrate Exec (no CREATE EXTENSION, never Fatal).
- (downloaded_at, artist) composite index; standalone downloaded_at index
  dropped. Native time scan replaces CAST AS CHAR; search uses ILIKE.
- Remove the favorite-column runtime probe; document 'pending' as transient.

Tests: Postgres integration test (real migrate + idempotency + ILIKE + composite
order + url_hash/content_hash concurrency) gated by OKFOLIO_TEST_POSTGRES_DSN and
a CI pgvector service container; sqlite kept for fast unit tests.

Refs #16


* Address review: derive category on conflict update, hash-index source_page

Two review findings from the prior PR:

- RecordDownload's ON CONFLICT DoUpdates map captured photo.Category before
  GORM runs BeforeSave. Callers (the scraper) construct photos without a
  Category and rely on the hook deriving it from SourcePage, so retry/conflict
  updates wrote an empty category and regrouped recovered downloads as
  "unknown". Derive the category up front via a shared resolveCategory helper
  used by both BeforeSave and the update assignment.

- SourcePage stored the full source page URL (scraper writes
  resolved.Source.URL) yet carried a plain btree (type:text;index) — the same
  unbounded-btree hazard removed from raw url: a long URL can exceed Postgres'
  ~2704-byte btree tuple limit and fail inserts. Drop the btree tag and create
  a bounded hash index (immune to that limit) in postMigratePostgres; drop any
  legacy btree of the GORM-default name first for idempotency.

Tests: sqlite unit test asserts a retry update persists the derived category;
Postgres integration test asserts source_page has a hash index and no btree,
and that a >2704-byte source_page inserts cleanly.


---------

+++ internal/api/api_test.go
@@ -583,7 +583,7 @@ func TestHandlePhotoDetailIncludesProvenanceAndFavorite(t *testing.T) {
 	}
 
 	var response struct {
-		ID        uint   `json:"id"`
+		ID        uint64 `json:"id"`
 		Source    string `json:"source"`
 		Provider  string `json:"provider"`
 		Category  string `json:"category"`
@@ -648,9 +648,9 @@ func TestHandleFavoritePersistsLocally(t *testing.T) {
 		t.Fatalf("Expected get favorite status 200, got %d", w.Code)
 	}
 	var response struct {
-		ID        uint `json:"id"`
-		Favorite  bool `json:"favorite"`
-		Available bool `json:"available"`
+		ID        uint64 `json:"id"`
+		Favorite  bool   `json:"favorite"`
+		Available bool   `json:"available"`
 	}
 	if err := json.NewDecoder(w.Body).Decode(&response); err != nil {
 		t.Fatalf("Failed to decode favorite status: %v", err)
